I fix build errors.
