+++ app/api/zora/rewards/route.tsx
@@ -1,30 +1,16 @@
 import { NextRequest } from 'next/server';
 import { Address } from 'viem';
 import trackEndpoint from '@/lib/stack/trackEndpoint';
-import getRewardsDepositLogs from '@/lib/zora/getRewardsDepositLogs';
-import getBlock from '@/lib/viem/getBlock';
-import formatBigIntValues from '@/lib/formatBigIntValues';
-import logsToStack from './logsToStack';
-import { EVENT_ZORA_REWARDS, INITIAL_BLOCK_RANGE } from '@/lib/consts';
-import { bulkTrack } from '@/lib/stack/bulkTrack';
-import getRewardsPoints from './getRewardsPoints';
+import { EVENT_ZORA_REWARDS } from '@/lib/consts';
+import getRewardsPoints from '@/lib/stack/getRewardsPoints';
+import indexNewRewards from '@/lib/indexNewRewards';
 
 export async function GET(request: NextRequest) {
   try {
     await trackEndpoint(EVENT_ZORA_REWARDS);
     const address = new URL(request.url).searchParams.get('address') as Address;
     const rewards = await getRewardsPoints(address);
-    const lastIndexedBlock = BigInt(rewards.events[0].metadata.blockNumber);
-    const startBlock = lastIndexedBlock || (await getBlock({ blockTag: 'earliest' })).number;
-    const latestBlock = await getBlock({ blockTag: 'latest' });
-    const rangeTo = startBlock + INITIAL_BLOCK_RANGE;
-    const isFinished = latestBlock.number <= rangeTo;
-    const toBlock = isFinished ? latestBlock.number : rangeTo;
-    const logs = await getRewardsDepositLogs(address, startBlock, toBlock);
-    if (logs?.length > 0) {
-      const trackingEvents = logsToStack(logs);
-      await bulkTrack(formatBigIntValues(trackingEvents));
-    }
+    const isFinished = await indexNewRewards(address, rewards);
     return Response.json({
       message: 'success',
       address,
