+++ lib/logsToStack.tsx
