+++ lib/stack/getRewardsPoints.tsx
